Changes as per review. Code now works on local and cloud env.
Fixed issue where tests code and tests were running successfully on local environment,
but the code was not executing on AWS. Was down to JSON serialization in the tests
and mock values being passed in where a string was expected.

Fixed issue where invoking another lambda was not working inside the VPC. This was
due to the Lambdas being inside public sg's and subnets, where infact it needed to be
in private alternatives so that it could access the internet.